grading/grader.py: initial import...
This code is in a bit of an odd state.  I initially wrote it for
Austin, but in an incomplete state, so it is peppered with
comments that direct him to fill things in, etc.

Since then I have made quick-and-dirty revisions when calculating
final grades at the end of each semester.

The code needs a good cleanup, a real user interface, and
generalization of many of the functions.  I am finally motivated
to actually do this, which is why I'm putting it in git.  Here goes!