* grading/grader.py: simplify interface to calculate_grade()
Each version of calculate_grade should now just accept a dictionary of
grades and return one.  Better separation of concerns now:  the
interface function (main_with_csv, etc.) just has to produce an
interable of entered grades, and the grade calculation function just
has to map a dict of entered grades to a dict of calculated grades in
whatever way is appropriate for each semester.