reporting/reports.py: fix a format string bug
In as_compact_text:

Python 3.4 was barfing when mean field in row_template had specified
precision, but the value passed to format was None.  The solution is to
only specify the precision when we know we have an actual value.

This fixes that issue, though I suspect there are similar issues with
min, max, etc. that just haven't surfaced.  And there is definitely a
similar issue in as_full_text.